LPD-6609 Moving EventsProcessorUtil and WebAppPool to portal-kernel
# breaking

## What portal-impl/src/com/liferay/portal/events/EventsProcessorUtil.java

Class has been moved to portal-kernel.

## Why

As this class contains core features, it is better to move it to portal-kernel to achieve accesibility from other classes.

## Alternatives

Use com.liferay.portal.kernel.events.EventsProcessorUtil instead.

----

# breaking

## What portal-impl/src/com/liferay/portal/util/WebAppPool.java

Class has been moved to portal-kernel.

## Why

As this class contains core features, it is better to move it to portal-kernel to achieve accesibility from other classes.

## Alternatives

Use com.liferay.portal.kernel.util.WebAppPool instead.


----
