+++ modules/apps/fragment/fragment-test/src/testIntegration/java/com/liferay/fragment/internal/struts/test/RenderFragmentEntryStrutsActionTest.java
@@ -8,7 +8,7 @@ package com.liferay.fragment.internal.struts.test;
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.petra.io.unsync.UnsyncStringWriter;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.events.EventsProcessorUtil;
+import com.liferay.portal.kernel.events.EventsProcessorUtil;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.RoleConstants;
 import com.liferay.portal.kernel.model.User;
