+++ modules/apps/layout/layout-portlets-web/src/main/java/com/liferay/layout/portlets/web/internal/display/context/LayoutPortletsDisplayContext.java
@@ -22,8 +22,8 @@ import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.kernel.util.WebAppPool;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.util.WebAppPool;
 
 import java.util.ArrayList;
 import java.util.Collection;
