+++ modules/apps/layout/layout-type-controller/layout-type-controller-panel/src/main/java/com/liferay/layout/type/controller/panel/taglib/internal/display/context/WidgetsTreeDisplayContext.java
@@ -26,11 +26,11 @@ import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.kernel.util.WebAppPool;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.kernel.util.comparator.PortletCategoryComparator;
 import com.liferay.portal.kernel.util.comparator.PortletTitleComparator;
 import com.liferay.portal.util.PortletCategoryUtil;
-import com.liferay.portal.util.WebAppPool;
 
 import java.util.ArrayList;
 import java.util.List;
