+++ modules/apps/roles/roles-admin-web/src/main/java/com/liferay/roles/admin/web/internal/display/context/EditRolePermissionsResourceDisplayContext.java
@@ -35,8 +35,8 @@ import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PortletKeys;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.kernel.util.WebAppPool;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.util.WebAppPool;
 import com.liferay.roles.admin.search.ResourceActionRowChecker;
 import com.liferay.roles.admin.web.internal.group.type.contributor.util.GroupTypeContributorUtil;
 import com.liferay.taglib.search.ResultRow;
